referly/requirements.txt
coreapi==2.1.1
dj-database-url==0.4.1
Django==1.10.4
django-extensions==1.7.5
django-webpack-loader==0.4.1
djangorestframework==3.5.3
gunicorn==19.6.0
httpie==0.9.9
itypes==1.1.0
psycopg2==2.6.2
Pygments==2.1.3
requests==2.12.3
six==1.10.0
uritemplate==3.0.0
whitenoise==3.2.2
